requirements.txt
torch
numpy
transformers
datasets
python-dotenv
huggingface_hub
openai